requirements.txt
albumentations==1.3.0
click==8.0.4
fastapi==0.89.1
gcsfs==2022.11.0
hydra-core==1.3.1
Markdown==3.4.1
matplotlib==3.5.1
numpy==1.22.3
opencv_python_headless==4.7.0.68
Pillow==9.4.0
protobuf==4.21.12
pytest==7.2.1
python-dotenv==0.21.0
segmentation_models_pytorch==0.3.2
setuptools==65.6.3
starlette==0.22.0
torch==1.13.1
torchvision==0.14.1
uvicorn
python-multipart